requirements.txt
fastapi
uvicorn
kiteconnect
gspread
google-auth
google-auth-oauthlib
python-dotenv
pytz
